Disconnect Scans module from Instruments module
